Enhance CI/CD configuration and improve integration test setup
