+++ .github/workflows/ReadSheet_build_and_test_on_main.yml
@@ -14,5 +14,10 @@ jobs:
     with:
       workdir: Frends.GoogleSheets.ReadSheet
       dotnet_version: 8.0.x
+      strict_analyzers: 'true'
+      env_var_name_1: GOOGLE_SERVICE_ACCOUNT_JSON_BASE64
+      env_var_name_2: GOOGLE_SHEET_ID
     secrets:
       badge_service_api_key: ${{ secrets.BADGE_SERVICE_API_KEY }}
+      env_var_value_1: ${{ secrets.GOOGLE_SERVICE_ACCOUNT_JSON_BASE64 }}
+      env_var_value_2: ${{ secrets.GOOGLE_SHEET_ID }}
