+++ .github/workflows/ReadSheet_release.yml
@@ -9,5 +9,6 @@ jobs:
     with:
       workdir: Frends.GoogleSheets.ReadSheet
       dotnet_version: 8.0.x
+      strict_analyzers: 'true'
     secrets:
       feed_api_key: ${{ secrets.TASKS_FEED_API_KEY }}
