+++ Frends.GoogleSheets.ReadSheet/Frends.GoogleSheets.ReadSheet.Tests/TestSetup.cs
@@ -1,0 +1,14 @@
+﻿using dotenv.net;
+using NUnit.Framework;
+
+namespace Frends.GoogleSheets.ReadSheet.Tests;
+
+[SetUpFixture]
+public class TestSetup
+{
+    [OneTimeSetUp]
+    public void GlobalSetup()
+    {
+        DotEnv.Fluent().WithEnvFiles(".env", ".env.development").Load();
+    }
+}
